+++ .github/ISSUE_TEMPLATE/bug_report.yaml
@@ -19,9 +19,9 @@ body:
       # Do not manually edit it.
       # Start components list
       - beyla.ebpf
-      - database_observability.mssql
       - database_observability.mysql
       - database_observability.postgres
+      - database_observability.sql_server
       - discovery.azure
       - discovery.consul
       - discovery.consulagent
