+++ .github/ISSUE_TEMPLATE/docs.yaml
@@ -22,9 +22,9 @@ body:
       # Do not manually edit it.
       # Start components list
       - beyla.ebpf
-      - database_observability.mssql
       - database_observability.mysql
       - database_observability.postgres
+      - database_observability.sql_server
       - discovery.azure
       - discovery.consul
       - discovery.consulagent
