+++ docs/sources/reference/compatibility/_index.md
@@ -45,9 +45,9 @@ The following components, grouped by namespace, _export_ Targets.
 {{< /collapse >}}
 
 {{< collapse title="database_observability" >}}
-- [database_observability.mssql](../components/database_observability/database_observability.mssql)
 - [database_observability.mysql](../components/database_observability/database_observability.mysql)
 - [database_observability.postgres](../components/database_observability/database_observability.postgres)
+- [database_observability.sql_server](../components/database_observability/database_observability.sql_server)
 {{< /collapse >}}
 
 {{< collapse title="discovery" >}}
@@ -130,9 +130,9 @@ The following components, grouped by namespace, _consume_ Targets.
 <!-- START GENERATED SECTION: CONSUMERS OF Targets -->
 
 {{< collapse title="database_observability" >}}
-- [database_observability.mssql](../components/database_observability/database_observability.mssql)
 - [database_observability.mysql](../components/database_observability/database_observability.mysql)
 - [database_observability.postgres](../components/database_observability/database_observability.postgres)
+- [database_observability.sql_server](../components/database_observability/database_observability.sql_server)
 {{< /collapse >}}
 
 {{< collapse title="discovery" >}}
@@ -263,9 +263,9 @@ The following components, grouped by namespace, _consume_ Loki `LogsReceiver`.
 <!-- START GENERATED SECTION: CONSUMERS OF Loki `LogsReceiver` -->
 
 {{< collapse title="database_observability" >}}
-- [database_observability.mssql](../components/database_observability/database_observability.mssql)
 - [database_observability.mysql](../components/database_observability/database_observability.mysql)
 - [database_observability.postgres](../components/database_observability/database_observability.postgres)
+- [database_observability.sql_server](../components/database_observability/database_observability.sql_server)
 {{< /collapse >}}
 
 {{< collapse title="faro" >}}
