+++ internal/component/all/all.go
@@ -3,9 +3,9 @@ package all
 
 import (
 	_ "github.com/grafana/alloy/internal/component/beyla/ebpf"                               // Import beyla.ebpf
-	_ "github.com/grafana/alloy/internal/component/database_observability/mssql"             // Import database_observability.mssql
 	_ "github.com/grafana/alloy/internal/component/database_observability/mysql"             // Import database_observability.mysql
 	_ "github.com/grafana/alloy/internal/component/database_observability/postgres"          // Import database_observability.postgres
+	_ "github.com/grafana/alloy/internal/component/database_observability/sql_server"        // Import database_observability.sql_server
 	_ "github.com/grafana/alloy/internal/component/discovery/aws"                            // Import discovery.aws.ec2 and discovery.aws.lightsail
 	_ "github.com/grafana/alloy/internal/component/discovery/azure"                          // Import discovery.azure
 	_ "github.com/grafana/alloy/internal/component/discovery/consul"                         // Import discovery.consul
