+++ internal/component/database_observability/sql_server/collector/connection_info.go
@@ -63,7 +63,7 @@ func (c *ConnectionInfo) Start(ctx context.Context) error {
 		providerRegion       = "unknown"
 		providerAccount      = "unknown"
 		dbInstanceIdentifier = "unknown"
-		engine               = "mssql"
+		engine               = "sql_server"
 	)
 
 	if c.CloudProvider != nil {
