+++ internal/component/database_observability/sql_server/collector/schema_details_test.go
